refactor(#57): got rid of wrappers around of Account and Main components
