+++ src/routing/routes/mainRoutes.tsx
@@ -2,7 +2,7 @@ import { Route } from 'react-router-dom';
 import AccountRouter from 'routing/routers/AccountRouter';
 import CreateMeetingRouter from 'routing/routers/CreateMeetingRouter';
 import JoinMeetingRouter from 'routing/routers/JoinMeetingRouter';
-import MainScreen from 'screens/main/MainScreen';
+import MainScreen from 'screens/main/MainScreen/MainScreen';
 import MeetingInfoScreen from 'screens/main/MeetingInfoScreen';
 
 import {
