+++ src/screens/account/AccountScreen/AccountScreen.tsx
@@ -1,21 +1,35 @@
-import React, { useCallback, useState } from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 
 import InfoItem from '@components/InfoItem/InfoItem';
+import useTgBackButton from 'hooks/useTgBackButton';
+import useTgMainButton from 'hooks/useTgMainButton';
 import { useNavigate } from 'react-router-dom';
 import {
   ACCOUNT_INTERFACE_LANGUAGES_PATH,
   ACCOUNT_LANGUAGES_PATH,
   ACCOUNT_LEVELS_PATH,
+  INSTANT_MAIN_PATH,
 } from 'routing/routing.constants';
 
-import styles from './Account.module.scss';
+import styles from './AccountScreen.module.scss';
 
-const Account = () => {
+const AccountScreen: React.FC = () => {
   const [language, setLanguage] = useState('English');
   const [level, setLevel] = useState('bginner');
   const [interfaceLanguage, setInterfaceLanguage] = useState('English');
 
   const navigate = useNavigate();
+
+  const { setBackButtonOnClick } = useTgBackButton(true);
+  useTgMainButton(false, false);
+
+  const handleBack = useCallback(() => {
+    navigate(INSTANT_MAIN_PATH);
+  }, [navigate]);
+
+  useEffect(() => {
+    setBackButtonOnClick(handleBack);
+  }, [handleBack, setBackButtonOnClick]);
 
   const openPracticeLanguages = useCallback(() => {
     navigate(ACCOUNT_LANGUAGES_PATH);
@@ -43,4 +57,4 @@ const Account = () => {
   );
 };
 
-export default Account;
+export default AccountScreen;
