+++ src/screens/main/MainScreen/MainScreen.module.scss
